requirements.txt
beautifulsoup4==4.9.3
Django==2.2.19
django-appconf==1.0.4
django-bootstrap4==3.0.1
django-taggit==0.22.2
django-user-accounts==3.0.4
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
gunicorn==20.0.4 
python-dotenv==0.13.0 